calibrate: don't add percentage margin on top of LUKS memory
kdumptool calibrate adds a margin (ADD_RESERVE_PCT = 30%) to the
calculated memory requirement.

In case of requirement for the LUKS/Argon2id PBKDF, the amount of memory
required is precisely specified and can be really large, so it makes no sense
to include this in the base for the +30% margin.

Move the addition of LUKS memory after ADD_RESERVE_PCT is added.

References: bsc#1229207